Added --dbfile <s> to cmcalibrate, which uses the GC content distribution in <s> instead of the default GC content distro. Also, replaced the default GC distro with RMARK-1's which is closer to normal with a mean of 0.5 than the previous default (which was Human chromosome 1). Replaced default
git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@2022 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/stats.h
@@ -30,7 +30,7 @@ void parallel_make_histogram (int *gc_count, int *partitions, int num_partitions
 			      int mpi_master_rank);
 #endif
 
-extern void GetDBInfo(ESL_SQFILE *sqfp, long *ret_N, int **ret_gc_ct);
+extern void GetDBInfo(ESL_SQFILE *sqfp, long *ret_N, double **ret_gc_ct);
 
 extern float e_to_score (float E, double *mu, double *lambda);
